# [Editor] Fix some performance issue happening on classification changes