@DavidKarlas 2018-10-09
I'm kinda skeptical this improves overall performance since this code was added explicitly to fix performance problems with rendering all lines on every change in file, maybe something changed in in mean time(Roslyn(not reporting whole file changed) or Editor improved something(doesn't re-render if nothing changed)).