@mkrueger 2018-10-11
@slluis: I proposed that chane :) - but wasn't sure about it. 

Because the @DavidKarlas comment is a valid argument but however it may not have a negative impact because these lines re render anyways and the highlighting is more than 'super slow' so checking may be worse. If that doesn't lead to any additional redraws it's way better.

However I don't have the profiling tools atm to verify.