@Therzok 2018-10-22
@monojenkins backport release-7.7